refactor: AuthContext 로직 분리 및 타입 모듈화
- AuthContext.ts: context 정의만 담당하도록 분리
- AuthProvider.tsx: 상태 및 로그인/로그아웃 로직 담당
- types.ts: User, AuthContextType 타입을 별도 파일로 추출
