+++ src/context/AuthContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { AuthContextType } from "./types";
+
+export const AuthContext = createContext<AuthContextType | null>(null);
