+++ src/context/AuthProvider.tsx
@@ -1,25 +1,11 @@
-import { createContext, useState } from "react";
+import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { getCurrentUser, postGithubLogin, postGithubLogout } from "../api/auth";
 import { LoadingFallback } from "../components/common";
 import { ROUTES } from "../constants";
 import { resetAuthState } from "../state/authState";
-
-interface User {
-  id: number;
-  name: string;
-  avatar: string;
-}
-
-export interface AuthContextType {
-  user: User | null;
-  isLogin: boolean;
-  setUser: (user: User | null) => void;
-  logout: () => Promise<void>;
-  loginWithCode: (code: string) => Promise<void>;
-}
-
-export const AuthContext = createContext<AuthContextType | null>(null);
+import { AuthContext } from "./AuthContext";
+import { User } from "./types";
 
 export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
   const [user, setUser] = useState<User | null>(null);
